+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.0-rc.4 : 06.09.2022
+
+- **Fixed**: Fixed missing validation errors
+- **Changed**: `Form.add_error()` now takes only `Tuple` as a `field` argument
+
 ## 1.0.0-rc.3 : 04.09.2022
 
 - **Fixed**: Removed validation of non-required fields if they are not present in the request